pyproject.toml
[project]
name = "my-langchain-project"
version = "0.1.0"
description = ""
authors = [
    {name = "kimdonghwan",email = "kehe12@nate.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain-openai (>=0.3.28,<0.4.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "openai (>=1.98.0,<2.0.0)",
    "streamlit (>=1.48.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pypdf (>=5.9.0,<6.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-chroma (>=0.2.5,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
